Move plan Approve/Cancel into a pinned bar above the composer
The proposed plan still renders inline in the chat scrollback, but its
Approve/Cancel decision moves into a new PlanApprovalTray pill in the
above-composer tray stack (alongside queue/goal), so the user can act
without scrolling back up to the plan card.
